refactor(agent-runner): split event_relay.rs into submodules to clear 600-line cap
event_relay.rs was 646 lines (46 over the workspace cap), blocking the
file-size CI check on every open PR. Extract two helper modules:

- event_relay/retry.rs  — post_with_retry + exponential-backoff constants + retry tests
- event_relay/dispatch.rs — flush_loop + flush_session (background drain loop)
- event_relay/mod.rs    — public API, types, spawn, buffer/notification tests

Public re-exports in lib.rs are unchanged; no call sites are affected.
All 25 agent-runner tests pass. Also fixes pre-existing clippy::cast_possible_wrap
in the two notification-gating tests (usize→i64 via i64::try_from).

Closes #610
